node.c: Fix theoretical endless iteration over named stream extents.
The type of 'num_extents' is a 32-bit unsigned value but 'k' was a
16-bit unsigned type, thus if the number of extents exceed 65535 the
condition 'k < num_extents' will always be true.

In practice this should never happen since we never iterate beyond
'value_size', which is a 16-bit unsigned value, but let's fix this to
silence static analyzer warnings and to make this safe in the case of
future changes.

+++ librefs/node.c
@@ -8082,7 +8082,7 @@ static int parse_attribute_named_stream_extent_value(
 
 	int err = 0;
 	u16 j = j_start;
-	u16 k = 0;
+	u32 k = 0;
 	u32 num_extents = 0;
 
 	/* 0x60 */
